+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com) and this project uses [semantic versioning](http://semver.org).
 
 ## [Unreleased]
+### Changed
+- Storage object get operations now also return the user ID if the owner is the root user.
+
 ### Fixed
 - Correctly display counters on console status page.
 - Correctly display friend names on console user details page.